Use nautilus actions instead of scripts menu
When using latest core with this commit:
http://git.qubes-os.org/gitweb/?p=marmarek/core.git;a=commitdiff;h=d12e532fc242a8e76f4861ad4db82debbe14315c

... the PDF convert action should be displayed in the "root" right-click menu,
thanks to the nautilus-actions.conf file that this commit brings in. This will
be default on Qubes R2 Beta 2. Also, the other actions, such as copy to other AppVM,
and open in Disposable VM have been rewritten to use nautilus actions.

Qubes R2 Beta 1 users will see the PDF convert in "Nautils-Actions" submenu, though.