client, server: Update to asyncio and new model
To prepare for a potential future move to another format besides RGB or
PNG, the intermediate files are now called Representations. Switching
over to another format simply consists of changing the name of a
parameter and adding/updating the convert command.

Intermediate files are now stored in a temporary directory for easier
cleanup.

Instead of a wrapper creating a client-vm subprocess which will execute
the client, we now have the client directly make the client-vm
subprocesses. This simplifies quite a bit as we can easily terminate
them if they act up, which becomes important since multiple files are
now supported.