server: Prevent CPU hogging by PNG tasks
PNG tasks were being enqueued too quickly, leaving no time for RGB conversions
or PNG deletions.  This meant that the server would create PNGs for every
single page of a PDF before any conversions started, which is clearly not
ideal.

After experimenting with limits on the number of PNGs created before forcing
the PNG creation task to join on the queue, I found that a limit of 1 gave the
best performance. Technically, it's a limit of 2 since we start a new task
before we await the previous one. In any case, the server is quite a bit faster
now and won't run out of space easily.